moving minimal pools back to their modules
