+++ src/Particle/tests/CMakeLists.txt
@@ -2,12 +2,17 @@
 #// This file is distributed under the University of Illinois/NCSA Open Source License.
 #// See LICENSE file in top directory for details.
 #//
-#// Copyright (c) 2021 QMCPACK developers.
+#// Copyright (c) 2025 QMCPACK developers.
 #//
 #// File developed by: Ye Luo, yeluo@anl.gov, Argonne National Laboratory
+#//                    Peter W. Doak, doakpw@ornl.gov, Oak Ridge National Lab
 #//
 #// File created by: Mark Dewing, markdewing@gmail.com, University of Illinois at Urbana-Champaign
 #//////////////////////////////////////////////////////////////////////////////////////
+
+add_library(minimal_particle_pool MinimalParticlePool.cpp)
+target_include_directories(minimal_particle_pool PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}")
+target_link_libraries(minimal_particle_pool PUBLIC qmcparticle)
 
 set(SRC_DIR particle)
 set(UTEST_EXE test_${SRC_DIR})
