+++ src/QMCHamiltonians/tests/CMakeLists.txt
@@ -2,13 +2,18 @@
 #// This file is distributed under the University of Illinois/NCSA Open Source License.
 #// See LICENSE file in top directory for details.
 #//
-#// Copyright (c) 2022 Jeongnim Kim and QMCPACK developers.
+#// Copyright (c) 2025 QMCPACK developers.
 #//
 #// File developed by: Mark Dewing, markdewing@gmail.com, University of Illinois at Urbana-Champaign
 #//                    Ye Luo, yeluo@anl.gov, Argonne National Laboratory
+#//                    Peter W. Doak, doakpw@ornl.gov, Oak Ridge National Lab
 #//
 #// File created by: Mark Dewing, markdewing@gmail.com, University of Illinois at Urbana-Champaign
 #//////////////////////////////////////////////////////////////////////////////////////
+
+add_library(minimal_hamiltonian_pool MinimalHamiltonianPool.cpp)
+target_include_directories(minimal_hamiltonian_pool PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}")
+target_link_libraries(minimal_hamiltonian_pool PUBLIC qmcham)
 
 set(SRC_DIR hamiltonian)
 set(UTEST_DIR ${CMAKE_CURRENT_BINARY_DIR})
