+++ src/QMCHamiltonians/tests/MinimalHamiltonianPool.cpp
