+++ src/QMCHamiltonians/tests/MinimalHamiltonianPool.h
