+++ src/QMCWaveFunctions/tests/MinimalWaveFunctionPool.cpp
